change in logic for calculating rotation angle to make coplanar - now using atan2 of (cross.length, dot product) - both of these terms check normal consistency first.  Then final rotation  = 180 - degrees
Next series of commits will remove debug logs, add more test cases, fixing naming, and do refactoring of large files.

+++ src/Unfold/ModelToGraph.cs
@@ -565,7 +565,7 @@ namespace Unfold
                     //weak code, shoould have a method for this - find edge that leads to
                     var edge = parent.Graph_Edges.Where(x => x.Head.Equals(child)).First();
 
-                    int nc = AlignPlanarFaces.CheckNormalConsistency(child.Face, parent.Face, edge.Real_Edge);
+                    double nc = AlignPlanarFaces.CheckNormalConsistency(child.Face, parent.Face, edge.Real_Edge);
                     Surface rotatedFace = AlignPlanarFaces.MakeGeometryCoPlanarAroundEdge(nc, child.UnfoldPolySurface, parent.Face, edge.Real_Edge) as Surface;
 
                     //at this point need to check if the rotated face has intersected with any other face that has been been
